pyproject.toml
[project]
name = "ai-chatbot-with-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix>=12.33.1",
    "chromadb>=1.4.1",
    "ddgs>=9.10.0",
    "gradio>=6.5.1",
    "openai-agents>=0.8.1",
    "openinference-instrumentation-openai-agents>=1.4.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
]
